requirements.txt
playwright
python-dotenv
pytz
nepse-api
pandas
requests
tqdm
